rename: /utils/heading.ts -> /utils/toc.ts
